+++ src/qt_ui/dynamics_dialog_ui.py
