+++ src/qt_ui/table_dialog_ui.py
